TOPS-102: Ansible Integration (#66)
* Change GitLab pipeline conditions

Run pipelines on all branches that have a subnet defined.
Destroy stage will not run on "prod" to prevent accidents.

* Run DNS role on all hosts

Role is designed to set up DNS server on appropriate guests, and
populate /etc/hosts for all guests.

* Add CUPS container

* Add initial top-level Ansible playbook

Currently handles basic configuration of all Ubuntu guests, as well as
DNS server/client setup, CUPS server, and rsyslog server.

* Add named config in DNS role

* Add Ansible execution to GitLab pipeline

* Add enviroment to all pipeline stages

I suspect this might be why the early stages aren't working with
the dev branch

* Install community.general Ansible collection

* Use the existing Ansible requirements file

* Run Ansible Galaxy again for collections

Collections are supposed to install during the general install command,
but that's not working for some reason.

* Rework root-level playbook

If you use include_playbook, it changes the scope to that playbook's
directory, so it can't find any of the roles and stuff.

* Add gateway to Terraform-created guests

* Fix gateway parameter

* Probably do gateway the right way

* Add SSH key for ubuntu-runner-1

The Ubuntu runner needs a way to SSH to the created guests for Ansible.

* Update public key for gitlab-runner

* Set Ansible remote_user to root

The containers don't have an ubuntu user; root only.

* Also use the Ansible inventory directory

The inventory directory has vault stuff and group vars, but Terraform
creates a hosts file for us in a different directory.  This should use
both.

* Fix typo for FreeIPA enrollment task

* Fix error in etc_hosts.j2

* Don't run FreeIPA enrollment if it's already done

* Correct name of DNS variable file

* Probably fix iterating over the DNS entries

* OK let's try a different method

* Finally fixed DNS variable formatting...

For some reason, doing the host entries like I was doing before wouldn't
allow anything to pick up the third variable in the dict.  I fixed this
by specifically formatting each host entry as a dict.

* Add rsyslog, cups to host inventory

Also add the rsyslog Terraform module.

* Identify DNS servers by IP instead of hostname

The Terraform-created hosts file uses IPs, not hostnames, so the DNS
servers weren't getting set up.

* Get Ansible Vault password from GitLab variables

* Fix rsyslog container IP

* Identify DNS servers in a different way

Regex wasn't working, but apparently I can use group_names to determine
if the current host is in the dns group.

* Force-enroll hosts in case they're being recreated

* Maybe fix DNS server template

* Linter cleanups

* Update BIND paths for Ubuntu

* Fix permissions on BIND files

* Rework inventory file format for hostnames

This should let the hostname-setting step actually work while retaining
the lack of dependency on DNS/hosts files for the runner.

* Fix getting the hostname from defined guests

* More linter fixes

* Even more linter fixes

* More DNS work

- Split up config files like Ubuntu expects
- Make sure directory ownership is correct

* MORE LINTER FIXES

* Update main.tf

* feat: add cloudflare DNS resolvers

In case google is having a bad day.

* Move /etc/hosts building to ubuntu role

* Fix all hosts becoming DNS servers

Oops.

Co-authored-by: claughinghouse <claughinghouse@users.noreply.github.com>